+++ src/things_mcp/tools_helpers/__init__.py
@@ -1,5 +1,13 @@
 """Tools package for Things 3 MCP server."""
 
 from .helpers import ToolsHelpers
+from .read_operations import ReadOperations
+from .write_operations import WriteOperations
+from .bulk_operations import BulkOperations
 
-__all__ = ['ToolsHelpers']
+__all__ = [
+    'ToolsHelpers',
+    'ReadOperations',
+    'WriteOperations',
+    'BulkOperations',
+]
